+++ src/main/java/br/com/catolica/nexus/Service/pacienteService.java
@@ -1,13 +1,12 @@
 package br.com.catolica.nexus.Service;
-
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import br.com.catolica.nexus.Model.Paciente;
 import br.com.catolica.nexus.repository.PacienteRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
 @Service
 public class PacienteService {
@@ -19,30 +18,34 @@ public class PacienteService {
         return repository.save(paciente);
     }
 
-    public List<Paciente> listarTodos() {
-        return repository.findAll();
+   
+    public Page<Paciente> listarTodos(Pageable pageable) {
+        return repository.findAll(pageable);
     }
+
+     //Metodo de buscar que lança exceção caso o paciente não seja encontrado (Julia)
 
     public Optional<Paciente> buscarPorId(String id) {
         return repository.findById(id);
     }
 
-    public Optional<Paciente> atualizar(String id, Paciente pacienteAtualizado) {
-        return repository.findById(id).map(paciente -> {
-            paciente.setNome(pacienteAtualizado.getNome());
-            paciente.setEmail(pacienteAtualizado.getEmail());
-            paciente.setTelefone(pacienteAtualizado.getTelefone());
-            paciente.setSenha(pacienteAtualizado.getSenha());
-            return repository.save(paciente);
-        });
-    }
 
-    public boolean deletar(String id) {
-        return repository.findById(id).map(p -> {
-            repository.deleteById(id);
-            return true;
-        }).orElse(false);
+     //Metodo de atualizar que lança exceção caso o paciente não seja encontrado (Julia)
+    public Paciente atualizar(String id, Paciente pacienteAtualizado) {
+        Paciente paciente = repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Paciente não encontrado com o id: " + id));
+        paciente.setNome(pacienteAtualizado.getNome());
+        paciente.setEmail(pacienteAtualizado.getEmail());
+        paciente.setTelefone(pacienteAtualizado.getTelefone());
+        paciente.setSenha(pacienteAtualizado.getSenha());
+        
+        return repository.save(paciente);
+    }
+ 
+    public void deletar(String id) {
+        if(!repository.existsById(id)){
+            throw new RuntimeException("Paciente não encontrado com o id: " + id);
+        }
+        repository.deleteById(id);
     }
 }
-
-
